+++ R/collapse.R
@@ -18,9 +18,14 @@
 #' )
 #' fct_count(partyid2)
 fct_collapse <- function(.f, ..., other_level = NULL, group_other = "DEPRECATED") {
+  f <- check_factor(.f)
 
   if (!missing(group_other)) {
-    warning("`group_other` is deprecated. Please use `other_level` instead")
+    lifecycle::deprecate_warn(
+      when = "0.5.0",
+      what = "fct_collapse(group_other)",
+      with = "fct_collapse(other_level)"
+    )
     if (isTRUE(group_other) && is.null(other_level)) {
       other_level <- "Other"
     }
@@ -29,8 +34,7 @@ fct_collapse <- function(.f, ..., other_level = NULL, group_other = "DEPRECATED"
   new <- rlang::list2(...)
   levs <- as.list(unlist(new, use.names = FALSE))
 
-  if (!is.null(other_level)){
-    f <- check_factor(.f)
+  if (!is.null(other_level)) {
     levels <- levels(f)
     new[[other_level]] <- levels[!levels %in% levs]
     levs <- c(levs, new[[other_level]])
@@ -39,8 +43,9 @@ fct_collapse <- function(.f, ..., other_level = NULL, group_other = "DEPRECATED"
   names(levs) <- names(new)[rep(seq_along(new), vapply(new, length, integer(1)))]
   out <- fct_recode(.f, !!!levs)
 
-  if (any(levels(out) == other_level)){
+  if (any(levels(out) == other_level)) {
     fct_relevel(out, other_level, after = Inf)
-  } else return(out)
-
+  } else {
+    out
+  }
 }