+++ R/data.R
@@ -3,7 +3,7 @@
 #' @source Downloaded from \url{https://gssdataexplorer.norc.org/}.
 #' @format
 #' \describe{
-#'  \item{year}{year of survey, 2000--2014}
+#'  \item{year}{year of survey, 2000--2014 (every other year)}
 #'  \item{age}{age. Maximum age truncated to 89.}
 #'  \item{marital}{marital status}
 #'  \item{race}{race}