+++ R/explicit_na.R
@@ -4,14 +4,17 @@
 #' appear in summaries and on plots.
 #'
 #' @param f A factor (or character vector).
-#' @param na_level Level to use for missing values: this is what NAs will be changed to.
+#' @param na_level Level to use for missing values: this is what `NA`s will be
+#'   changed to.
 #' @export
 #' @examples
 #' f1 <- factor(c("a", "a", NA, NA, "a", "b", NA, "c", "a", "c", "b"))
 #' fct_count(f1)
+#' table(is.na(f1))
 #'
 #' f2 <- fct_explicit_na(f1)
 #' fct_count(f2)
+#' table(is.na(f2))
 fct_explicit_na <- function(f, na_level = "(Missing)") {
   f <- check_factor(f)