+++ R/unique.R
@@ -11,4 +11,3 @@ fct_unique <- function(f) {
   factor(levels(f), levels(f), exclude = NULL, ordered = is.ordered(f))
 }
 
-