+++ tests/testthat/test-match.R
@@ -1,5 +1,3 @@
-context("test-fct_match.R")
-
 test_that("equivalent to %in% when levels present", {
   f <- factor(c("a", "b", "c", NA))
   expect_equal(fct_match(f, "a"), f %in% "a")
@@ -8,5 +6,5 @@ test_that("equivalent to %in% when levels present", {
 
 test_that("error when levels are missing", {
   f <- factor(c("a", "b", "c"))
-  expect_error(fct_match(f, "d"), "not present")
+  expect_snapshot(fct_match(f, "d"), error = TRUE)
 })