+++ tests/testthat/test-relabel.R
@@ -1,27 +1,17 @@
-context("fct_relabel")
-
 test_that("identity", {
   f1 <- factor(c("a", "b"))
 
   expect_identical(fct_relabel(f1, identity), f1)
 })
 
-test_that("error if not function", {
+test_that("gives useful errors", {
   f1 <- factor("a")
 
-  expect_error(fct_relabel(f1, 1), "function")
-})
-
-test_that("error if level not character", {
-  f1 <- factor("a")
-
-  expect_error(fct_relabel(f1, function(x) 1), "character")
-})
-
-test_that("error if level has different length", {
-  f1 <- factor(letters)
-
-  expect_error(fct_relabel(f1, function(x) x[-1]), "expected 26.*got 25")
+  expect_snapshot(error = TRUE, {
+    fct_relabel(f1, 1)
+    fct_relabel(f1, function(x) 1)
+    fct_relabel(f1, function(x) x[-1])
+  })
 })
 
 test_that("total collapse", {
@@ -41,14 +31,14 @@ test_that("formulas are coerced to functions", {
   f1 <- factor(letters)
 
   expect_identical(
-    fct_relabel(f1, ~paste0(.x, ".")),
+    fct_relabel(f1, ~ paste0(.x, ".")),
     factor(paste0(letters, "."))
   )
 })
 
 test_that("string input is coerced to a factor", {
   expect_identical(
-    fct_relabel(LETTERS[1:2], .fun=function(x) x),
+    fct_relabel(LETTERS[1:2], function(x) x),
     factor(LETTERS[1:2])
   )
 })