+++ tests/testthat/test-utils.R
@@ -1,10 +1,15 @@
-context("test-utils.R")
-
 test_that("check_factor() fails when needed", {
-  expect_error(check_factor(NA), "factor")
+  expect_equal(check_factor("x"), factor("x"))
+  expect_equal(check_factor(factor("x")), factor("x"))
+
+  expect_snapshot(error = TRUE, {
+    check_factor(NA)
+  })
 })
 
-test_that("check_factor_list() fails when needed", {
-  expect_error(check_factor_list(1), "list")
-  expect_error(check_factor_list(list(1)), "factor")
+test_that("check_factor_list() checks its inputs", {
+  expect_snapshot(error = TRUE, {
+    check_factor_list(1)
+    check_factor_list(list(1))
+  })
 })