+++ NEWS.md
@@ -2,6 +2,13 @@
 * `fct_reordern()` is a new function to order based on an arbitrary number of
   values (@billdenney, #16)
 
+* New `fct()` which works like `factor()` but errors if values of `x`
+  are not included in the levels specification (#299)
+
+* `first2()` and `last2()` now ignore missing values in both `x` and `y` (#303).
+
+* Error messages are more informative.
+
 # forcats 0.5.1
 
 * Re-license as MIT (#277).